+++ rasa_core_sdk/events.py
@@ -144,10 +144,11 @@ def AgentUttered(text=None, data=None, timestamp=None):
 
 
 # noinspection PyPep8Naming
-def Form(name, timestamp=None):
+def Form(name, trigger_message=None, timestamp=None):
     return {
         "event": "form",
         "name": name,
+        "trigger_message": trigger_message,
         "timestamp": timestamp,
     }