+++ CHANGELOG.rst
@@ -14,6 +14,10 @@ This project adheres to `Semantic Versioning`_ starting with version 0.11.0.
 Added
 -----
 - add optional `validate_{slot}` methods to `FormAction`
+- ``FormAction.from_trigger_intent`` allows slot extraction from message
+  triggering the FormAction
+- ``Tracker.active_form`` now includes ``trigger_message`` attribute to allow
+  access to message triggering the form
 
 Removed
 -------