+++ CHANGELOG.rst
@@ -14,6 +14,8 @@ This project adheres to `Semantic Versioning`_ starting with version 0.11.0.
 Added
 -----
 - add optional `validate_{slot}` methods to `FormAction`
+- Function to get latest input channel from the tracker with
+  ``tracker.get_latest_input_channel()``
 
 Removed
 -------