+++ CHANGELOG.rst
@@ -4,6 +4,26 @@ Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning`_ starting with version 0.11.0.
 
+.. _master-release:
+
+[Unreleased 0.13.0.aX] - `master`_
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+.. note:: This version is not yet released and is under active development.
+
+Added
+-----
+- add warning in case of mismatched version of rasa_core and rasa_core_sdk
+
+Changed
+-------
+
+Fixed
+-----
+
+Removed
+-------
+
 [0.12.2] - 2019-02-17
 ^^^^^^^^^^^^^^^^^^^^^