+++ CHANGELOG.rst
@@ -24,6 +24,7 @@ Changed
 -------
 - ``self._deactivate()`` method from the ``FormAction`` class has been
   renamed to ``self.deactivate()``
+- changed endpoint function so that it is now accessible with Python as well
 
 Fixed
 -----